Move test connection and server into util
